restructuring of core.py
